+++ docs/conf.py
@@ -22,7 +22,7 @@ copyright = '2021, Bo Peng'
 author = 'Bo Peng'
 
 # The full version, including alpha/beta/rc tags
-release = '1.3'
+release = '1.4'
 
 # -- General configuration ---------------------------------------------------