+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+1.4
+
+- Fix a bug in calculating nfft and noverlap in pipeline.analyze_performance()
+
 1.3
 
 - Add function double_nanmad_flag and method DataObj.get_double_nanmad_flag, which do MAD flagging twice to deal with
@@ -11,7 +15,7 @@
 - Change the criteria for flagging time series outlier to flag different chop phase separately
 - Fix a bug in rebuilding time stamp
 - Fix a bug in desnaking to deal with the case that ref_pix is in pix_flag_list
-- Fix a bug in FigFlux.imshow_flag which doesn't accept empty list
+- Fix a bug in FigFlux.imshow_flag() which doesn't accept empty list
 
 1.2